chore: moved OnfidoInitialize into VerifyIdentity
